+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/CreatingHoneySlime.java
@@ -3,6 +3,7 @@ package com.telepathicgrunt.the_bumblezone.entities;
 import com.telepathicgrunt.the_bumblezone.entities.mobs.HoneySlimeEntity;
 import com.telepathicgrunt.the_bumblezone.modinit.BzCriterias;
 import com.telepathicgrunt.the_bumblezone.modinit.BzEntities;
+import com.telepathicgrunt.the_bumblezone.modinit.BzSounds;
 import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
 import com.telepathicgrunt.the_bumblezone.utils.GeneralUtils;
 import net.minecraft.core.BlockPos;
@@ -55,7 +56,7 @@ public class CreatingHoneySlime {
                     playerEntity.getX(),
                     playerEntity.getY(),
                     playerEntity.getZ(),
-                    SoundEvents.SLIME_BLOCK_STEP,
+                    honeySlimeMob.isBaby() ? BzSounds.HONEY_SLIME_SQUISH_SMALL : BzSounds.HONEY_SLIME_SQUISH,
                     SoundSource.NEUTRAL,
                     1.0F,
                     1.0F);