+++ src/main/java/com/telepathicgrunt/the_bumblezone/modinit/BzSounds.java
@@ -17,6 +17,18 @@ public class BzSounds {
     public static final SoundEvent STINGER_SPEAR_HIT_GROUND = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "item.the_bumblezone.stinger_spear.hit_ground"));
     public static final SoundEvent STINGER_SPEAR_RETURN = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "item.the_bumblezone.stinger_spear.return"));
     public static final SoundEvent STINGER_SPEAR_THROW = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "item.the_bumblezone.stinger_spear.throw"));
+    public static final SoundEvent WASHING_RESIDUES = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "block.the_bumblezone.washing_honey_residues"));
+    public static final SoundEvent HONEY_SLIME_ATTACK = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.attack"));
+    public static final SoundEvent HONEY_SLIME_HURT = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.hurt"));
+    public static final SoundEvent HONEY_SLIME_DEATH = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.death"));
+    public static final SoundEvent HONEY_SLIME_SQUISH = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.squish"));
+    public static final SoundEvent HONEY_SLIME_JUMP = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.jump"));
+    public static final SoundEvent HONEY_SLIME_HURT_SMALL = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.hurt_small"));
+    public static final SoundEvent HONEY_SLIME_DEATH_SMALL = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.death_small"));
+    public static final SoundEvent HONEY_SLIME_SQUISH_SMALL = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.squish_small"));
+    public static final SoundEvent HONEY_SLIME_JUMP_SMALL = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.honey_slime.jump_small"));
+    public static final SoundEvent POLLEN_PUFF_THROW = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "entity.the_bumblezone.pollen_puff.throw"));
+    public static final SoundEvent SUGAR_WATER_DRINK = new SoundEvent(new ResourceLocation(Bumblezone.MODID, "item.the_bumblezone.sugar_water_bottle.drink"));
 
     public static void registerSounds() {
         Registry.register(Registry.SOUND_EVENT, ANGERED_BEES.getLocation(), ANGERED_BEES);
@@ -30,5 +42,17 @@ public class BzSounds {
         Registry.register(Registry.SOUND_EVENT, STINGER_SPEAR_HIT_GROUND.getLocation(), STINGER_SPEAR_HIT_GROUND);
         Registry.register(Registry.SOUND_EVENT, STINGER_SPEAR_RETURN.getLocation(), STINGER_SPEAR_RETURN);
         Registry.register(Registry.SOUND_EVENT, STINGER_SPEAR_THROW.getLocation(), STINGER_SPEAR_THROW);
+        Registry.register(Registry.SOUND_EVENT, WASHING_RESIDUES.getLocation(), WASHING_RESIDUES);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_ATTACK.getLocation(), HONEY_SLIME_ATTACK);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_HURT.getLocation(), HONEY_SLIME_HURT);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_DEATH.getLocation(), HONEY_SLIME_DEATH);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_SQUISH.getLocation(), HONEY_SLIME_SQUISH);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_JUMP.getLocation(), HONEY_SLIME_JUMP);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_HURT_SMALL.getLocation(), HONEY_SLIME_HURT_SMALL);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_DEATH_SMALL.getLocation(), HONEY_SLIME_DEATH_SMALL);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_SQUISH_SMALL.getLocation(), HONEY_SLIME_SQUISH_SMALL);
+        Registry.register(Registry.SOUND_EVENT, HONEY_SLIME_JUMP_SMALL.getLocation(), HONEY_SLIME_JUMP_SMALL);
+        Registry.register(Registry.SOUND_EVENT, POLLEN_PUFF_THROW.getLocation(), POLLEN_PUFF_THROW);
+        Registry.register(Registry.SOUND_EVENT, SUGAR_WATER_DRINK.getLocation(), SUGAR_WATER_DRINK);
     }
 }